signforge-local/src/signforge/training/validate.py
"""
Validation logic for SignForge training.
"""

from __future__ import annotations
from typing import Dict, List, Any

import torch
from torch.utils.data import DataLoader
from PIL import Image

from signforge.core.logging import get_logger
from signforge.ml.eval.metrics import MetricsCalculator

logger = get_logger(__name__)


class Validator:
    """Handles validation and image generation during training."""
    
    def __init__(self, device: str = "cuda") -> None:
        self.device = device
        self.metrics_calc = MetricsCalculator()

    @torch.no_grad()
    def validate_epoch(
        self, 
        model: torch.nn.Module, 
        dataloader: DataLoader
    ) -> Dict[str, float]:
        """Run validation on the provided dataloader."""
        model.eval()
        total_loss = 0.0
        
        for batch in dataloader:
            # Placeholder for real validation forward pass
            loss = 0.1 
            total_loss += loss
            
        avg_loss = total_loss / len(dataloader)
        return {"val_loss": avg_loss}

    def generate_samples(
        self,
        model: Any,
        prompts: List[str],
        num_samples: int = 4
    ) -> List[Image.Image]:
        """Generate sample images to visually inspect progress."""
        # This would interface with the inference pipeline/SignForgePipeline
        # but using the current training weights
        logger.info("generating_validation_samples", count=len(prompts))
        
        # Mocking sample generation
        samples = []
        for _ in range(min(len(prompts), num_samples)):
            samples.append(Image.new("RGB", (512, 512), (200, 200, 200)))
            
        return samples

    def evaluate_quality(self, images: List[Image.Image]) -> Dict[str, float]:
        """Calculate quality metrics for generated samples."""
        if not images:
            return {}
            
        batch_metrics = self.metrics_calc.batch_calculate(images)
        # Flatten summary stats
        return {
            "avg_sharpness": batch_metrics.get("sharpness", {}).get("mean", 0.0),
            "avg_technical_score": batch_metrics.get("technical_score", {}).get("mean", 0.0),
        }
